requirements.txt
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.2
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0.13
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-reversion==3.0.3
elasticsearch==2.4.0
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.2.0
jdcal==1.4
moviepy==1.0.0
mysql==0.0.2
mysqlclient==1.4.2.post1
numpy==1.16.1
odfpy==1.4.0
openpyxl==2.6.0
Pillow==5.4.1
proglog==0.1.9
PyMySQL==0.9.3
pytz==2018.9
PyYAML==5.1
redis==3.2.0
requests==2.21.0
six==1.10.0
tablib==0.12.1
tqdm==4.31.1
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
